🐛 엣지 런타임에서 사용 불가능한 패키지 사용으로 인한 핸들러 이분화
