+++ src/app/api/recipe/video/[id]/route.ts
@@ -1,21 +1,17 @@
 import { NextRequest, NextResponse } from "next/server";
 import { YoutubeTranscript } from "youtube-transcript";
 
-import { OpenAIService } from "@/service/openAI";
-
 interface Context {
   params: { id: string };
 }
 
-interface script {
+export interface Script {
   offset: number;
   text: string;
 }
 
-export const runtime = "edge";
-
 export async function GET(_: NextRequest, { params: { id } }: Context) {
-  const YTtranscript: script[] = await YoutubeTranscript.fetchTranscript(
+  const YTtranscript: Script[] = await YoutubeTranscript.fetchTranscript(
     id,
   ).then((res) => {
     return res.map(({ text, offset }) => ({
@@ -24,16 +20,5 @@ export async function GET(_: NextRequest, { params: { id } }: Context) {
     }));
   });
 
-  try {
-    const message: string = YTtranscript.map((s) => s.text).join("\n");
-    console.log(message);
-    const openai = OpenAIService.getInstance();
-    const response = await openai.getChatResponse(message);
-    console.log(response);
-    const content = response.choices[0].message.content;
-    return NextResponse.json(content);
-  } catch (error) {
-    console.error(error);
-    return NextResponse.error();
-  }
+  return NextResponse.json(YTtranscript);
 }
